kernel: remove some redundant casts

+++ src/compiler.c
@@ -3159,7 +3159,7 @@ CVar CompElmListLev (
 {
     CVar                lists;          /* lists                           */
     CVar                pos;            /* position                        */
-    Int                 level;          /* level                           */
+    UInt                level;          /* level                           */
 
     /* compile the lists expression                                        */
     lists = CompExpr(READ_EXPR(expr, 0));
@@ -3169,7 +3169,7 @@ CVar CompElmListLev (
     CompCheckIntSmallPos( pos );
 
     /* get the level                                                       */
-    level = (Int)(READ_EXPR(expr, 2));
+    level = READ_EXPR(expr, 2);
 
     /* emit the code to select the elements from several lists (to <lists>)*/
     Emit( "ElmListLevel( %c, %c, %d );\n", lists, pos, level );
@@ -3191,7 +3191,7 @@ CVar CompElmsListLev (
 {
     CVar                lists;          /* lists                           */
     CVar                poss;           /* positions                       */
-    Int                 level;          /* level                           */
+    UInt                level;          /* level                           */
 
     /* compile the lists expression                                        */
     lists = CompExpr(READ_EXPR(expr, 0));
@@ -3200,7 +3200,7 @@ CVar CompElmsListLev (
     poss = CompExpr(READ_EXPR(expr, 1));
 
     /* get the level                                                       */
-    level = (Int)(READ_EXPR(expr, 2));
+    level = READ_EXPR(expr, 2);
 
     /* emit the code to select the elements from several lists (to <lists>)*/
     Emit( "ElmsListLevelCheck( %c, %c, %d );\n", lists, poss, level );
@@ -3267,7 +3267,7 @@ CVar CompElmRecName (
     record = CompExpr(READ_EXPR(expr, 0));
 
     /* get the name (stored immediately in the expression)                 */
-    rnam = (UInt)(READ_EXPR(expr, 1));
+    rnam = READ_EXPR(expr, 1);
     CompSetUseRNam( rnam, COMP_USE_RNAM_ID );
 
     /* emit the code to select the element of the record                   */
@@ -3337,7 +3337,7 @@ CVar CompIsbRecName (
     record = CompExpr(READ_EXPR(expr, 0));
 
     /* get the name (stored immediately in the expression)                 */
-    rnam = (UInt)(READ_EXPR(expr, 1));
+    rnam = READ_EXPR(expr, 1);
     CompSetUseRNam( rnam, COMP_USE_RNAM_ID );
 
     /* emit the code to test the element                                   */
@@ -3531,7 +3531,7 @@ CVar CompElmComObjName (
     record = CompExpr(READ_EXPR(expr, 0));
 
     /* get the name (stored immediately in the expression)                 */
-    rnam = (UInt)(READ_EXPR(expr, 1));
+    rnam = READ_EXPR(expr, 1);
     CompSetUseRNam( rnam, COMP_USE_RNAM_ID );
 
     /* emit the code to select the element of the record                   */
@@ -3618,7 +3618,7 @@ CVar CompIsbComObjName (
     record = CompExpr(READ_EXPR(expr, 0));
 
     /* get the name (stored immediately in the expression)                 */
-    rnam = (UInt)(READ_EXPR(expr, 1));
+    rnam = READ_EXPR(expr, 1);
     CompSetUseRNam( rnam, COMP_USE_RNAM_ID );
 
     /* emit the code to test the element                                   */
@@ -4152,11 +4152,11 @@ void CompFor (
             vart = 'm';
         }
         else if (TNUM_EXPR(READ_STAT(stat, 0)) == T_REF_HVAR) {
-            var = (UInt)(READ_EXPR(READ_STAT(stat, 0), 0));
+            var = READ_EXPR(READ_STAT(stat, 0), 0);
             vart = 'h';
         }
         else /* if ( TNUM_EXPR( READ_STAT(stat, 0) ) == T_REF_GVAR ) */ {
-            var = (UInt)(READ_EXPR(READ_STAT(stat, 0), 0));
+            var = READ_EXPR(READ_STAT(stat, 0), 0);
             CompSetUseGVar( var, COMP_USE_GVAR_ID );
             vart = 'g';
         }
@@ -4711,7 +4711,7 @@ void CompAssListLev (
     CVar                lists;          /* lists                           */
     CVar                pos;            /* position                        */
     CVar                rhss;           /* right hand sides                */
-    Int                 level;          /* level                           */
+    UInt                level;          /* level                           */
 
     /* print a comment                                                     */
     if ( CompPass == 2 ) {
@@ -4729,7 +4729,7 @@ void CompAssListLev (
     rhss = CompExpr(READ_STAT(stat, 2));
 
     /* get the level                                                       */
-    level = (Int)(READ_STAT(stat, 3));
+    level = READ_STAT(stat, 3);
 
     /* emit the code                                                       */
     Emit( "AssListLevel( %c, %c, %c, %d );\n", lists, pos, rhss, level );
@@ -4751,7 +4751,7 @@ void CompAsssListLev (
     CVar                lists;          /* list                            */
     CVar                poss;           /* positions                       */
     CVar                rhss;           /* right hand sides                */
-    Int                 level;          /* level                           */
+    UInt                level;          /* level                           */
 
     /* print a comment                                                     */
     if ( CompPass == 2 ) {
@@ -4768,7 +4768,7 @@ void CompAsssListLev (
     rhss = CompExpr(READ_STAT(stat, 2));
 
     /* get the level                                                       */
-    level = (Int)(READ_STAT(stat, 3));
+    level = READ_STAT(stat, 3);
 
     /* emit the code                                                       */
     Emit( "AsssListLevelCheck( %c, %c, %c, %d );\n",
@@ -4832,7 +4832,7 @@ void CompAssRecName (
     record = CompExpr(READ_STAT(stat, 0));
 
     /* get the name (stored immediately in the statement)                  */
-    rnam = (UInt)(READ_STAT(stat, 1));
+    rnam = READ_STAT(stat, 1);
     CompSetUseRNam( rnam, COMP_USE_RNAM_ID );
 
     /* compile the right hand side                                         */
@@ -4901,7 +4901,7 @@ void CompUnbRecName (
     record = CompExpr(READ_STAT(stat, 0));
 
     /* get the name (stored immediately in the statement)                  */
-    rnam = (UInt)(READ_STAT(stat, 1));
+    rnam = READ_STAT(stat, 1);
     CompSetUseRNam( rnam, COMP_USE_RNAM_ID );
 
     /* emit the code for the assignment                                    */
@@ -5098,7 +5098,7 @@ void CompAssComObjName (
     record = CompExpr(READ_STAT(stat, 0));
 
     /* get the name (stored immediately in the statement)                  */
-    rnam = (UInt)(READ_STAT(stat, 1));
+    rnam = READ_STAT(stat, 1);
     CompSetUseRNam( rnam, COMP_USE_RNAM_ID );
 
     /* compile the right hand side                                         */
@@ -5183,7 +5183,7 @@ void CompUnbComObjName (
     record = CompExpr(READ_STAT(stat, 0));
 
     /* get the name (stored immediately in the statement)                  */
-    rnam = (UInt)(READ_STAT(stat, 1));
+    rnam = READ_STAT(stat, 1);
     CompSetUseRNam( rnam, COMP_USE_RNAM_ID );
 
     /* emit the code for the assignment                                    */
